add self deletion prevention

+++ app/Core/Organization/Http/Controllers/API/V1/Admin/UserAPIController.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Core\Organization\Http\Controllers\API\V1\Admin;
 
 use App\Contracts\Controller\BaseAPIController;
+use App\Core\Organization\Exceptions\UserException;
 use App\Core\Organization\Http\Requests\V1\Admin\StoreUserRequest;
 use App\Core\Organization\Http\Requests\V1\Admin\UpdateUserRequest;
 use App\Core\Organization\Http\Transformers\V1\Admin\UserTransformer;
@@ -12,6 +13,7 @@ use App\Core\Organization\Mappers\UserMapper;
 use App\Core\Organization\Repositories\UserRepository;
 use App\Core\Organization\Services\UserService;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
 
 class UserAPIController extends BaseAPIController
 {
@@ -50,9 +52,14 @@ class UserAPIController extends BaseAPIController
         );
     }
 
-    public function destroy(int $id): JsonResponse
+    public function destroy(int $id, Request $request): JsonResponse
     {
         $user = $this->repository->findOrFail($id);
+
+        if ($request->user()->id === $user->id) {
+            throw UserException::selfDeletionPrevention();
+        }
+
         $this->service->delete($user);
 
         return $this->response->success([], 'کاربر مورد نظر حذف شد.');
